Crate volta_migrate

Provides types for modeling the current state of the Volta directory and for migrating between versions

A new layout should be represented by its own struct (as in the existing v0 or v1 modules) Migrations between types should be represented by TryFrom implementations between the layout types (see v1.rs for examples)

NOTE: Since the layout file is written once the migration is complete, all migration implementations need to be aware that they may be partially applied (if something fails in the process) and should be able to re-start gracefully from an interrupted migration

Modules

detect 🔒

empty 🔒

v0 🔒

v1 🔒

v2 🔒

v3 🔒

v4 🔒

Macros

detect_tagged 🔒

Macro to simplify the boilerplate associated with detecting a tagged state.

Enums

MigrationState 🔒

Represents the state of the Volta directory at every point in the migration process

Functions

detect_and_migrate 🔒

run_migration