volta_migrate

Module v2

Structs

V2

Represents a V2 Volta Layout (used by Volta v0.7.3 and above)

Functions

clear_default_npm 🔒

Clear npm from the default platform.json file if it is set to the same value as that bundled with Node

remove_npm_version_from_node_image_dir 🔒

Move a single node image up a directory, if it currently has the npm version in its path

shift_node_images 🔒

Move all Node images up one directory, removing the default npm version directory