Function volta_migrate::v2::shift_node_images

fn shift_node_images(old_home: &VoltaHome, new_home: &VoltaHome) -> Fallible<()>

Move all Node images up one directory, removing the default npm version directory

In the V1 layout, we kept all node images in /<node_version>/<npm_version>/, however we will be storing custom npm versions in a separate image directory, so there is no need to maintain the bundled npm version in the file structure any more. This also will make it slightly easier to access the Node image, as we no longer will need to look up the bundled npm version every time.