volta_migrate

Module v3

Modules

config 🔒

Structs

V3

Represents a V3 Volta layout (used by Volta v0.9.0 and above)

Functions

get_installed_packages 🔒

Determine a list of all installed packages that are using the legacy package config

is_migrated_config 🔒

Determine if a package has already been migrated by attempting to read the V3 PackageConfig

migrate_packages 🔒

migrate_single_package 🔒

Migrate a single package to the new workflow