Function volta_migrate::v3::migrate_single_package

fn migrate_single_package(
    config: LegacyPackageConfig, 
    session: &mut Session
) -> Fallible<()>

Migrate a single package to the new workflow

Note: This relies on the package install logic in volta_core. If that logic changes, then the end result may not be a valid V3 layout any more, and this migration will need to be updated. Specifically, the invariants we rely on are:

Package image directory is in the same location

Package config files are in the same location and the same format

Binary config files are in the same location and the same format

If any of those are violated, this migration may be invalid and need to be reworked / scrapped